+++ ChangeLog.md
@@ -5,6 +5,7 @@
 **Released: WiP**
 
 - The bookmark search command palette now sorts the bookmarks.
+  ([#2](https://github.com/davep/hike/pull/2))
 - The suggester for the command line now prefers newer history entries over
   older. ([#3](https://github.com/davep/hike/pull/3))