+++ ChangeLog.md
@@ -8,7 +8,7 @@
   ([#2](https://github.com/davep/hike/pull/2))
 - The suggester for the command line now prefers newer history entries over
   older. ([#3](https://github.com/davep/hike/pull/3))
-- Add the ability to move the command line to the top of the screen.
+- Added the ability to move the command line to the top of the screen.
   ([#5](https://github.com/davep/hike/pull/5))
 
 ## v0.1.0