+++ ChangeLog.md
@@ -11,7 +11,9 @@
 - Added the ability to move the command line to the top of the screen.
   ([#5](https://github.com/davep/hike/pull/5))
 - Added the ability to copy the current location to the clipboard.
+  ([#9](https://github.com/davep/hike/pull/9))
 - Added the ability to copy the currently-viewed Markdown to the clipboard.
+  ([#9](https://github.com/davep/hike/pull/9))
 
 ## v0.1.0